requirements.txt
numpy 
matplotlib
pandas
scikit-learn
open3d
opencv-python
pandas==1.0.5
scipy==1.5.0
scikit_image==0.16.2
openpyxl==3.0.4
cloudpickle==1.5.0
numpy==1.18.5
smplx==0.1.14
joblib==0.16.0
numexpr==2.7.1
Shapely==1.7.1
tqdm==4.47.0
psutil==5.7.0
torchvision==0.7.0
pyrender==0.1.43
requests==2.24.0
matplotlib==3.2.2
caffe2==0.8.1
cityscapesscripts==2.0.0
dataclasses==0.7
filterpy==1.4.5
fvcore==0.1.1.post20200716
heatmap==2.2.1
motmetrics==1.2.0
moviepy==1.0.3
onnx==1.7.0
paf==0.2.1
Pillow==9.0.0
pycocotools==2.0.2
recommonmark==0.6.0
scikit_learn==0.23.2
skimage==0.0
sphinx_rtd_theme==0.5.0
tabulate==0.8.7
tensorboardX==2.1
termcolor==1.1.0
